changed the time of applying the rules

+++ gameLife/src/settings.hpp
@@ -2,10 +2,10 @@
 #define SETTINGS_HPP
 struct settings
 {
-	unsigned short WINDOW_WIDTH = 800;
-	unsigned short WINDOW_HEIGHT = 800;
+	int WINDOW_WIDTH = 800;
+	int WINDOW_HEIGHT = 800;
 
-	unsigned short BOARD_WIDTH = 10;
-	unsigned short BOARD_HEIGHT = 10;
+	int BOARD_WIDTH = 10;
+	int BOARD_HEIGHT = 10;
 } settings;
 #endif // SETTINGS_HPP
